Add a manually-triggered GitHub workflow (#25)
* Add a manually-triggered GitHub workflow

This workflow can be used to test `action-upterm` on a specified runner
pool. This comes in handy to verify that it works on a given platform or
even in a given container.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* Fix limit-access-to-actor default value

The default was 'auto' but that wasn't in the options list, which would
cause the workflow to fail when triggered without explicitly selecting
a value.

---------

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Co-authored-by: Owen Ou <o@owenou.com>